requirements.txt
ailia==1.4.0.0
ailia-tokenizer==1.3.1.0
contourpy==1.3.0
cycler==0.12.1
dlib==19.24.6
filterpy==1.4.5
fonttools==4.54.1
ftfy==6.2.3
h5py==3.12.1
imageio==2.35.1
kiwisolver==1.4.7
lapx==0.5.10.post1
lazy_loader==0.4
llvmlite==0.43.0
matplotlib==3.9.2
networkx==3.3
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyMatting==1.1.12
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.9.20
tqdm==4.66.5
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
streamlit
